requirements.txt
pandas
selenium
beautifulsoup4
python-dotenv
openpyxl
tabulate
requests